test: move tests to public folder and fix waitFor
